+++ src/redirects/redirects-list.js
@@ -23,3 +23,4 @@ export { DidomiLoader } from './didomi-loader';
 export { Prebid } from './prebid';
 export { prebidAds } from './prebid-ads';
 export { NaverWcslog } from './naver-wcslog';
+export { FreewheelAdManager } from './freewheel-admanager';