+++ src/scriptlets/freewheel-admanager.ts
@@ -1,0 +1,17 @@
+import { FreewheelAdManager, FreewheelAdManagerNames } from '../redirects/freewheel-admanager';
+
+/**
+ * @scriptlet freewheel-admanager
+ *
+ * @description
+ * Mocks the FreeWheel Ad Manager.
+ *
+ * ### Examples
+ *
+ * ```adblock
+ * example.org#%#//scriptlet('freewheel-admanager')
+ * ```
+ *
+ * @added unknown.
+ */
+export { FreewheelAdManager, FreewheelAdManagerNames };