+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ### Added
 
+- `freewheel-admanager` redirect resource and scriptlet [#401].
 - `getCreativeId` method to `Ad` class in `google-ima3` redirect [#515].
 
 ### Changed
@@ -33,6 +34,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
   `debug-current-inline-script` and `log-on-stack-trace` scriptlets [#513].
 
 [Unreleased]: https://github.com/AdguardTeam/Scriptlets/compare/v2.2.16...HEAD
+[#401]: https://github.com/AdguardTeam/Scriptlets/issues/401
 [#513]: https://github.com/AdguardTeam/Scriptlets/issues/513
 [#515]: https://github.com/AdguardTeam/Scriptlets/issues/515
 [#545]: https://github.com/AdguardTeam/Scriptlets/issues/545